Split refresh from update
Previous to this commit the data and the UI was refreshed at the same time. If a user typed a key, or resized the window, all of the data queries would be reissued. This would lead to slow refreshes and high cpu usage.

now the data update schedule is separate from the ui refresh schedule.