Fix coord calculations
A few off-by-one errors were plauging the coord calcs in mapper.py. For example, [90,180] would fail to draw anything. This commit changes the lat_long_to_x_y func to better calculate the actual location of a coord. As a result, the compass border was also slightly off, so this commit fixed that as well.