Fix issue with sending the root of a pool
Prior to this commit, if you tried to shadow the root of a pool you'd
get a failure. This was due to the cleanup logic that would remove
the remote destination of your send. If you were sending the root, you'd
end up removing the target dataset. This commit makes it so the destination
is re-created.