Add --insecure to port forward request
It appears that PIA uses self-signed cigs for their port forwarding REST service, which was causing cURL to die a dishonerable death when attempting to fetch a forwarded port. This commit adds the '--insecure' option to cURL so it ignores cert errors.

NOTE: this is a potential security risk. This means that a man-in-the-middle could pose as PIA and recieve your request for a forwarded port. For me personally this is an acceptable risk: OpenVPN will make sure the VPN provider I'm connecting to is actually PIA, if some malicious 3rd party wants to hand me a not-actually-forwarded port, then so be it, my torrents will be slower for it, but they'll know fuck-all about what I'm actually sending over the wire.