Add safemv command
safemv is a safe way to move a file to a directory. If the file
doesn't exist in the destination, it's moved. If the file does
exist but with a different hash, the file is moved to the directory
but with a new filename such that it doesn't collied. If the file
does exist and has the same hash, then the file being moved is deleted.